requirements.txt
flask==1.0.2
flask-cors==3.0.6
flask-sqlalchemy==2.3.1
geoalchemy2==0.4.2
psycopg2==2.7.4
marshmallow==2.15.3
apispec==0.38.0
gunicorn==19.8.1
